Update game reviews to add sorting, filters and proper length constraints

+++ database/migrations/2025_02_19_183410_create_game_reviews_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('game_reviews', function (Blueprint $table) {
             $table->id();
-            $table->integer('rating');
+            $table->tinyInteger('rating');
             $table->text('review');
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->foreignId('game_id')->constrained()->onDelete('cascade');
